+++ .ci/ubuntu/gha-setup.sh
@@ -7,7 +7,7 @@ set -o xtrace
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 readonly script_dir
 echo "[INFO] script_dir: '$script_dir'"
-readonly rabbitmq_image=rabbitmq:4.1.0-alpine
+readonly rabbitmq_image=rabbitmq:4.1.0-beta.4-management-alpine
 
 
 readonly docker_name_prefix='rabbitmq-amqp-python-client'
